+++ lib/sub_page/company/my_company/my_company_widget.dart
@@ -499,9 +499,8 @@ class _MyCompanyWidgetState extends State<MyCompanyWidget>
                                         'ReviewList',
                                         queryParameters: {
                                           'companyID': serializeParam(
-                                            myCompanyVarItem.companyID
-                                                .toDouble(),
-                                            ParamType.double,
+                                            myCompanyVarItem.companyID,
+                                            ParamType.int,
                                           ),
                                         }.withoutNulls,
                                         extra: <String, dynamic>{
